% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{aggregate_from_individual_tidy}
\alias{aggregate_from_individual_tidy}
\title{Aggregate Individual Fish Records to Summarized Format (tidyverse version)}
\usage{
aggregate_from_individual_tidy(
  data,
  treatment_col = "tmt",
  replicate_col = "tank",
  score_col = "score",
  total_col = "total"
)
}
\arguments{
\item{data}{A data frame containing individual fish records}

\item{treatment_col}{Name of the column containing treatment groups (default: "tmt")}

\item{replicate_col}{Name of the column containing tank/replicate IDs (default: "tank")}

\item{score_col}{Name of the column containing scores (default: "score")}

\item{total_col}{Name of the column to be created for total counts (default: "total")}
}
\value{
A data frame with aggregated counts per replicate
}
\description{
Aggregate Individual Fish Records to Summarized Format (tidyverse version)
}
