% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_dunnett}
\alias{broom_dunnett}
\title{Standardize Dunnett Test Results}
\usage{
broom_dunnett(
  x,
  method = c("Dunnett_multcomp", "Dunnett_DescTools", "Dunnett_PMCMRplus"),
  ...
)
}
\arguments{
\item{x}{A formula object specifying the response variable and the factor for which the test is to be performed,
or an object of class 'aov' or 'lm'.}

\item{method}{Character string specifying which implementation of Dunnett test to use:
"Dunnett_multcomp" (default) for multcomp implementation,
"Dunnett_DescTools" for DescTools implementation, or
"Dunnett_PMCMRplus" for PMCMRplus implementation, do not use this one in general as it does not return complete information.}

\item{...}{Additional arguments passed to the underlying test functions.}
}
\value{
A data frame with standardized test results containing:
\item{comparison}{The comparison being made}
\item{estimate}{The estimated difference between groups}
\item{p.value}{The p-value for the test}
\item{conf.low}{Lower bound of the confidence interval}
\item{conf.high}{Upper bound of the confidence interval}
\item{method}{The method used for the test}
}
\description{
This function creates a standardized output from different Dunnett test implementations.
}
