% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R, R/zzz_aliases.R
\name{calcSteepnessOverlap}
\alias{calcSteepnessOverlap}
\alias{calc_steepness_overlap}
\title{Steepness and overlap calcuation}
\usage{
calcSteepnessOverlap(
  mod = NULL,
  obj = NULL,
  trend = "Decrease",
  CI = "inv",
  ...
)

calc_steepness_overlap(
  mod = NULL,
  obj = NULL,
  trend = "Decrease",
  CI = "inv",
  ...
)
}
\arguments{
\item{mod}{fitted object from drm model fitting}

\item{obj}{Calculated ED 10, 20, 50 object if available.
mod should set to be NULL in this case}

\item{trend}{"Decrease" or "Increase"}

\item{CI}{method to calculate CI, either "inv" (bmd) or "delta" or "fls".}

\item{...}{other parameters that will be passed into ED.plus}
}
\value{
a table with certainty of potection level and steepness of the models
}
\description{
Steepness and overlap calcuation
}
\examples{
\dontrun{
 datTn<- subset(oecd201,Time==72)
 mod <- drm(Yield~Concentration,data=datTn,fct=LL.3())
 calcSteepnessOverlap(mod = mod, trend = "Decrease")
}
}
