% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOEC.R
\name{calculate_noec_rstatix}
\alias{calculate_noec_rstatix}
\title{Calculate NOEC Using Many-to-one Pairwise Tests}
\usage{
calculate_noec_rstatix(
  data,
  response,
  dose,
  control = "0",
  test = c("t.test", "wilcox.test"),
  p_adjust_method = "holm",
  alternative = "two.sided",
  alpha = 0.05
)
}
\arguments{
\item{data}{A data frame containing the dose response data}

\item{response}{The name of the response variable (unquoted)}

\item{dose}{The name of the dose variable (unquoted)}

\item{control}{The level of the dose variable to be used as control}

\item{test}{The statistical test to use: "t.test" or "wilcox.test"}

\item{p_adjust_method}{Method for p-value adjustment for multiple comparisons}

\item{alternative}{Direction of the alternative hypothesis: "two.sided", "greater", or "less"}

\item{alpha}{Significance level (default: 0.05)}
}
\value{
A list containing the NOEC value and the full test results
}
\description{
This function calculates the No Observed Effect Concentration (NOEC) from dose response data
using pairwise comparison tests from the rstatix package.
}
