% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{collembola_juveniles}
\alias{collembola_juveniles}
\title{Fake data from collembola juveniles}
\format{
collembola_juveniles in wide format
}
\usage{
collembola_juveniles
}
\description{
Fake data from collembola juveniles
}
\keyword{datasets}
