% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{compare_phi_methods}
\alias{compare_phi_methods}
\title{Compare phi estimation methods (FIXED)}
\usage{
compare_phi_methods(successes, totals, doses, scoring = "doses")
}
\arguments{
\item{successes}{Vector of success counts}

\item{totals}{Vector of total trials}

\item{doses}{Vector of dose levels}

\item{scoring}{Scoring method}
}
\description{
Compare phi estimation methods (FIXED)
}
