% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantal.R
\name{compare_to_control_fisher}
\alias{compare_to_control_fisher}
\title{Perform Fisher's Exact Test Comparing Each Level to Control}
\usage{
compare_to_control_fisher(
  data,
  factor_col,
  success_col,
  failure_col,
  control_level = NULL,
  alternative = "two.sided",
  conf.level = 0.95
)
}
\arguments{
\item{data}{A data frame containing the data}

\item{factor_col}{String name of the column containing the factor (e.g., "dose")}

\item{success_col}{String name of the column containing success counts (e.g., "alive")}

\item{failure_col}{String name of the column containing failure counts (e.g., "dead")}

\item{control_level}{The level to use as control (default: first level in the data)}

\item{alternative}{Direction of the alternative hypothesis ("two.sided", "less", or "greater")}

\item{conf.level}{Confidence level for the returned confidence interval}
}
\value{
A data frame with factor levels and corresponding p-values
}
\description{
This function performs Fisher's exact test comparing each level of a factor
to a control level, using counts of success and failure.
}
\examples{
# Generate example data
set.seed(123)
doses <- c(0, 0.1, 1, 10, 50, 100)
reps <- 4
data <- data.frame(
  dose = rep(doses, each = reps),
  replicate = rep(1:reps, times = length(doses))
)
data$alive <- sapply(data$dose, function(x) rbinom(1, 10, prob = exp(-0.02 * x)))
data$dead <- 10 - data$alive

# Run Fisher's exact test
compare_to_control_fisher(data, "dose", "alive", "dead", control_level = 0)

}
