% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{complete_trend_analysis}
\alias{complete_trend_analysis}
\title{Perform complete overdispersion and trend analysis (CORRECTED)}
\usage{
complete_trend_analysis(
  successes,
  totals,
  doses,
  scoring = c("doses", "ranks", "log_doses", "equal_spaced"),
  alternative = c("two.sided", "greater", "less"),
  phi_method = c("trend_adjusted", "simple")
)
}
\arguments{
\item{successes}{Vector of success counts}

\item{totals}{Vector of total trials}

\item{doses}{Vector of dose levels}

\item{scoring}{Scoring method to use consistently}

\item{alternative}{Direction of trend test}

\item{phi_method}{Method for phi estimation when using Rao-Scott}
}
\value{
List with Tarone and step-down results
}
\description{
Perform complete overdispersion and trend analysis (CORRECTED)
}
