% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{comprehensive_phi_comparison}
\alias{comprehensive_phi_comparison}
\title{Compare phi estimates across different scoring methods and estimation methods}
\usage{
comprehensive_phi_comparison(successes, totals, doses)
}
\arguments{
\item{successes}{Vector of success counts}

\item{totals}{Vector of total trials}

\item{doses}{Vector of dose levels}
}
\description{
Compare phi estimates across different scoring methods and estimation methods
}
