% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{compute_mdd_williams}
\alias{compute_mdd_williams}
\title{Calculate MDD\% for a Williams Test Result}
\usage{
compute_mdd_williams(williams_obj, data, formula)
}
\arguments{
\item{williams_obj}{The tibble result from broom_williams.}

\item{data}{The original dataframe used for the test.}

\item{formula}{The formula used for the test, e.g., Response ~ Dose.}
}
\value{
A tibble with Dose, MDD, and MDD_pct.
}
\description{
Calculate MDD\% for a Williams Test Result
}
