% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS.R
\name{convert2Score}
\alias{convert2Score}
\title{Convert Values to Scores}
\usage{
convert2Score(Dvec)
}
\arguments{
\item{Dvec}{A numeric vector to be converted.}
}
\value{
A numeric vector with non-positive values set to NA.
}
\description{
Converts any non-positive numbers to NA.
}
