% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_binom.R
\name{create_summary_table}
\alias{create_summary_table}
\title{Create a summary table for manuscript/report}
\usage{
create_summary_table(comparison_results)
}
\arguments{
\item{comparison_results}{Output from compare_tarone_scoring}
}
\description{
Create a summary table for manuscript/report
}
