% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{DixonQ}
\alias{DixonQ}
\title{Dixon's outlier test critical Q table}
\format{
An object of class \code{data.frame} with 14 rows and 2 columns.
}
\usage{
DixonQ
}
\description{
Dixon's outlier test critical Q table
}
\references{
DIXON, W. J. (1950) Analysis of extreme values. Ann. Math. Stat. 21, 488-506.
DEAN, R. B., DIXON, W. J. (1951) Simplified statistics for small numbers of observation. Anal. Chem. 23, 636-638.
}
\author{
Zhenglei Gao
}
\keyword{datasets}
