% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R, R/zzz_aliases.R
\name{dose.p.glmmPQL}
\alias{dose.p.glmmPQL}
\alias{dose_p_glmm_pql}
\title{Calculate Dose Probability for GLMM using PQL}
\usage{
dose.p.glmmPQL(obj, cf = 1:2, p = 0.5)

dose_p_glmm_pql(obj, cf = 1:2, p = 0.5)
}
\arguments{
\item{obj}{A fitted GLMM object of class "glmmPQL".}

\item{cf}{A numeric vector specifying the coefficients to be used for the calculation (default is 1:2).}

\item{p}{A numeric value representing the probability level (default is 0.5).}
}
\value{
A structured object of class "glm.dose" containing the estimated dose probability and its standard error.
}
\description{
This function calculates the dose probability for a generalized linear mixed model (GLMM)
using Penalized Quasi-Likelihood (PQL).
}
\examples{
# Assuming `glmm_model` is a fitted glmmPQL model
## library(MASS)
glmm_model <- MASS::glmmPQL(yt ~ dose,random= ~1 | Obs,family= quasibinomial(link="logit"),
data=pvi_example)
dose_result <- dose.p.glmmPQL(glmm_model)
print(dose_result)

}
