% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{williamsTestLookUpTable}
\alias{williamsTestLookUpTable}
\alias{test_cases_data}
\alias{dat_shallow}
\alias{dat_steep}
\alias{dat_medium}
\alias{dat_noED50}
\title{Williams Test Lookup Table}
\format{
An object of class \code{data.frame} with 119 rows and 19 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 768 rows and 17 columns.

An object of class \code{data.frame} with 50 rows and 2 columns.

An object of class \code{data.frame} with 24 rows and 2 columns.

An object of class \code{data.frame} with 24 rows and 2 columns.

An object of class \code{data.frame} with 24 rows and 2 columns.
}
\usage{
williamsTestLookUpTable

test_cases_data

dat_shallow

dat_steep

dat_medium

dat_noED50
}
\description{
Williams Test Lookup Table

test_cases_data for validation purpose

example data with shallow dose-response

example data with steep dose-response

example data with medium dose-response

example data with no ED 50 response
}
\keyword{datasets}
