% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R, R/zzz_aliases.R
\name{ED.plus}
\alias{ED.plus}
\alias{ED.ZG}
\alias{ed_plus}
\title{Calculating ED following the regulatory ED definition.}
\usage{
\method{ED}{plus}(
  object,
  respLev,
  maxEff = TRUE,
  trend = "Increase",
  range = "Percentage",
  CI = c("delta", "inv", "bmd-inv"),
  ...
)

\method{ED}{ZG}(...)

ed_plus(
  object,
  respLev,
  maxEff = TRUE,
  trend = "Increase",
  range = "Percentage",
  CI = c("delta", "inv", "bmd-inv"),
  ...
)
}
\arguments{
\item{object}{fitted model using drc::drm}

\item{respLev}{response level, the x of ECx.}

\item{maxEff}{Not used for now. maximum effect.}

\item{trend}{Increase of Decrease, whether the dose response is decreasing or increasing compared control.}

\item{range}{using Percentage, not used for now.}

\item{CI}{methods to calculate the confidence intervals, either "inv", "bmd-inv",(bmd) or "delta" or "fls".}

\item{...}{other parameters for ED function}
}
\value{
Back calculated regulatory ECx
}
\description{
Calculating ED following the regulatory ED definition.

Deprecated helper function for ED calculation
}
\details{
Due to old ECxHelper development where ED.plus is defined as ED.ZG
}
\examples{
\dontrun{
data("dat_medium")
dat_medium <- dat_medium |> mutate(Treatment=factor(Dose,levels=unique(Dose)))
dat_medium$Response[dat_medium$Response < 0] <- 0
mod <- drm(Response~Dose,data=dat_medium,fct=LL.3())
fctList <- list(LN.4(),LL.4(),W1.3(),LL2.2())
res <- mselect.plus(mod,fctList = fctList )
modList <- res$modList
mod <-modList[[1]]
edres <- ED.plus(mod,c(10,50),trend="Decrease")
}
}
\seealso{
\code{\link[=ED.plus]{ED.plus()}} for the public version of usage
}
\keyword{Deprecated}
