% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{ED.ZG}
\alias{ED.ZG}
\title{Deprecated helper function for ED calculation}
\usage{
\method{ED}{ZG}(
  object,
  respLev,
  maxEff = TRUE,
  trend = "Increase",
  range = "Percentage",
  CI = c("delta", "inv", "bmd-inv"),
  ...
)
}
\description{
Deprecated helper function for ED calculation
}
\details{
Due to old ECxHelper development where ED.plus is defined as ED.ZG
}
\seealso{
\code{\link[=ED.plus]{ED.plus()}} for the public version of usage
}
\keyword{Deprecated}
