% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{estimate_phi_with_scoring}
\alias{estimate_phi_with_scoring}
\title{Estimate overdispersion parameter using consistent scoring (CORRECTED)}
\usage{
estimate_phi_with_scoring(
  successes,
  totals,
  doses,
  scoring,
  method = c("trend_adjusted", "simple")
)
}
\arguments{
\item{successes}{Vector of success counts}

\item{totals}{Vector of total trials}

\item{doses}{Vector of dose levels}

\item{scoring}{Scoring method to use}

\item{method}{Method for phi estimation: "trend_adjusted" or "simple"}
}
\description{
Estimate overdispersion parameter using consistent scoring (CORRECTED)
}
