% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{exampleHistData}
\alias{exampleHistData}
\title{exampleHistData from StatCharrms}
\format{
An object of class \code{data.frame} with 671 rows and 56 columns.
}
\usage{
exampleHistData
}
\description{
exampleHistData from StatCharrms
}
\keyword{datasets}
