% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{expand_to_individual_tidy}
\alias{expand_to_individual_tidy}
\title{Expand Aggregated Fish Data to Individual Records (tidyverse version)}
\usage{
expand_to_individual_tidy(
  data,
  treatment_col = "tmt",
  replicate_col = "tank",
  score_prefix = "S",
  total_col = "total"
)
}
\arguments{
\item{data}{A data frame containing aggregated fish data}

\item{treatment_col}{Name of the column containing treatment groups (default: "tmt")}

\item{replicate_col}{Name of the column containing tank/replicate IDs (default: "tank")}

\item{score_prefix}{Prefix used for score columns (default: "S")}

\item{total_col}{Name of the column containing total counts (default: "total")}
}
\value{
A data frame with individual fish records
}
\description{
Expand Aggregated Fish Data to Individual Records (tidyverse version)
}
