% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/williams_JT.R, R/zzz_aliases.R
\name{getComparison}
\alias{getComparison}
\alias{get_comparison}
\title{Function to get the comparison groups for Williams test or other.}
\usage{
getComparison(object, test = c("trend", "many-to-one"))

get_comparison(object, test = c("trend", "many-to-one"))
}
\arguments{
\item{object}{test result from different type of tests}

\item{test}{the test used, either a trend test or a many-to-one test}
}
\description{
Function to get the comparison groups for Williams test or other.
}
