% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R, R/zzz_aliases.R
\name{getwilliamRes}
\alias{getwilliamRes}
\alias{get_william_res}
\title{get from william res accept/reject}
\usage{
getwilliamRes(william, n = NULL)

get_william_res(william, n = NULL)
}
\arguments{
\item{william}{william test results}

\item{n}{number of hypotheses to be tested. description}
}
\value{
a vector of accept and rejects.
}
\description{
get from william res accept/reject
}
\examples{
## Example from Sachs (1997, p. 402)
x <- c(106, 114, 116, 127, 145,
       110, 125, 143, 148, 151,
       136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("0", "I", "II")

## Williams Test
res <- PMCMRplus::williamsTest(x ~ g)
getwilliamRes(res)
}
