% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsr_tsk.R
\docType{data}
\name{hamilton}
\alias{hamilton}
\title{This is hamilton data}
\format{
A list containing ten data frames: dr1a, dr1b, dr1c,
dr1d, dr1e, dr4a, dr4b, dr4c, dr4d, dr4e,
}
\source{
Hamilton, 1977.
}
\usage{
hamilton
}
\description{
Example dose-response data given in Hamilton (1977).
Note that, as per Hamilton (1978), the confidence intervals
given in Hamilton (1977) for these data sets are incorrect.
}
\references{
\url{data_blah.com}
}
\keyword{data}
