% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{invlogxp}
\alias{invlogxp}
\title{inverse transformation of logxp}
\usage{
invlogxp(x, a)
}
\arguments{
\item{x}{a numeric vector}

\item{a}{a numeric vector}
}
\value{
inverse log of x and a
}
\description{
inverse transformation of logxp
}
\examples{
invlogxp(3, 1)
}
