% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_response_simulation.R
\name{log_message}
\alias{log_message}
\title{Adding detailed logging}
\usage{
log_message(message, log_file = "simulation_log.txt")
}
\arguments{
\item{message}{message to be written}

\item{log_file}{file to be appended to}
}
\value{
log_file with series of messages if specified in the simulation
}
\description{
Adding detailed logging
}
