% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{logxp}
\alias{logxp}
\title{Helper function for scaling x-axis}
\usage{
logxp(x, a)
}
\arguments{
\item{x}{a vector of numeric}

\item{a}{a vector of nummeric}
}
\value{
log(x+a)
}
\description{
Helper function for scaling x-axis
}
\examples{
logxp(0,1)
}
