% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{metaldata}
\alias{metaldata}
\title{Data from heavy metal mixture experiments (drcData)}
\format{
An object of class \code{data.frame} with 543 rows and 3 columns.
}
\usage{
metaldata
}
\description{
Data are from a study of the response of the cyanobacterial
self-luminescent metallothionein-based whole-cell biosensor
Synechoccocus elongatus PCC 7942 pBG2120 to binary mixtures of
6 heavy metals (Zn, Cu, Cd, Ag, Co and Hg).
}
\keyword{datasets}
