% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R, R/zzz_aliases.R
\name{mselect.ED}
\alias{mselect.ED}
\alias{mselect_ed}
\title{ECx calculation together with normalized width proposed by EFSA SO.}
\usage{
mselect.ED(modList, respLev = c(10, 20, 50), trend = "Decrease", ...)

mselect_ed(modList, respLev = c(10, 20, 50), trend = "Decrease", ...)
}
\arguments{
\item{modList}{list of models}

\item{respLev}{response levels or ECx for example, 10, 20, 50}

\item{trend}{"Decrease" or "Incrrease".}

\item{...}{optional to pass to ED function.}
}
\value{
ED result table
}
\description{
ECx calculation together with normalized width proposed by EFSA SO.
}
