% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{mselect.ZG}
\alias{mselect.ZG}
\title{Deprecated helper function for mselect}
\usage{
mselect.ZG(
  object = NULL,
  fctList = NULL,
  nested = FALSE,
  sorted = c("IC", "Res var", "Lack of fit", "no"),
  linreg = FALSE,
  icfct = AIC,
  respCol = "effect",
  doseCol = "dose",
  data = NULL,
  type = "continuous",
  additionalReliability = c("EFSA")
)
}
\description{
Deprecated helper function for mselect
}
\details{
Due to old ECxHelper development where mselect.plus is defined as mselect.ZG
}
\keyword{Deprecated}
