% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{NTA_Ar_ext_mortality_expected}
\alias{NTA_Ar_ext_mortality_expected}
\title{Expected outcome of pre-processing NTA_Ar_ext mortality}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 7 columns.
}
\usage{
NTA_Ar_ext_mortality_expected
}
\description{
Expected outcome of pre-processing NTA_Ar_ext mortality
}
\keyword{datasets}
