% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{NTP_example}
\alias{NTP_example}
\title{NTP_example data}
\format{
NTP example data with corrected and uncorrected emergence, survival, reduction.
}
\usage{
NTP_example
}
\description{
NTP_example data
}
\keyword{datasets}
