% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{NTP_example_rate}
\alias{NTP_example_rate}
\title{Expected processed data from the NTP_example}
\format{
data rame with corrected and uncorrected emergence, survival, reduction.
}
\usage{
NTP_example_rate
}
\description{
Expected processed data from the NTP_example
}
\keyword{datasets}
