% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{NTP_example_rate_expected}
\alias{NTP_example_rate_expected}
\title{Expected outcome of NTP}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 8 columns.
}
\usage{
NTP_example_rate_expected
}
\description{
Expected outcome of NTP
}
\keyword{datasets}
