% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{oecd201}
\alias{oecd201}
\title{An example dataset from study type OECD 201}
\format{
A data frame with 112 rows and 14 columns
}
\description{
An example dataset from study type OECD 201
}
\details{
The data is originally provided by Ecotox colleague Andreas and was preprocessed so that GrowthRate is included.
}
\references{
OECD Test No. 201: Freshwater Alga and Cyanobacteria, Growth Inhibition Test \url{https://www.oecd.org/en/publications/test-no-201-alga-growth-inhibition-test_9789264069923-en.html}
}
\author{
BCS
}
\keyword{datasets}
