% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_binom.R
\name{perform_groupwise_tarone}
\alias{perform_groupwise_tarone}
\title{Perform Tarone test within each dose group}
\usage{
perform_groupwise_tarone(data)
}
\arguments{
\item{data}{Data frame with successes, totals, doses}
}
\description{
Perform Tarone test within each dose group
}
