% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{plot.dunnett_test_result}
\alias{plot.dunnett_test_result}
\title{Plot method for dunnett_test_result objects}
\usage{
\method{plot}{dunnett_test_result}(x, ...)
}
\arguments{
\item{x}{A dunnett_test_result object}

\item{...}{Additional arguments passed to plot methods}
}
\description{
Plot method for dunnett_test_result objects
}
