% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{plot.edList}
\alias{plot.edList}
\title{Plot the ECx estimation and confidence intervals from the list of models.}
\usage{
\method{plot}{edList}(edList, fctNames, ...)
}
\arguments{
\item{edList}{ECx lists from fitted set of models}

\item{fctNames}{fct function names}

\item{...}{additional parameters}
}
\value{
a ggplot object
}
\description{
Plot the ECx estimation and confidence intervals from the list of models.
}
