% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{plot.modList}
\alias{plot.modList}
\title{Plot a list of models together.}
\usage{
\method{plot}{modList}(
  x,
  respLev = NULL,
  data = NULL,
  xmin,
  xmax,
  scale = c("logx", "logy", "logxy", "orig"),
  npts = 100,
  plot_respLev = FALSE,
  xbreaks = NULL,
  ymin = NULL,
  ymax = NULL,
  ...
)
}
\arguments{
\item{x}{a list of drc models with modList class.}

\item{respLev}{to be calculated ECx levels, for example, c(10,20,50).}

\item{data}{data used to fit the set of models}

\item{xmin}{minimum value of xaxis, should be greater than 0 to
fit the log scale.}

\item{xmax}{maximum value of xaxis, can be missing.}

\item{scale}{could be "logx", "logy", "logxy", "orig".}

\item{npts}{number of points used for plotting the prediction
confidence bands}

\item{plot_respLev}{logical, whether to add the estimated ECx with CIs}

\item{xbreaks}{breaks in x-axis labeling}

\item{ymin}{minimum of y}

\item{ymax}{max of y}

\item{...}{addition parameters to be passed into plotting functions or
user input when the use does not want to use the
data provided in the modList}
}
\value{
a ggplot object
}
\description{
Plot a list of models together.
}
\examples{
\dontrun{
data("dat_medium")
dat_medium <- dat_medium |> mutate(Treatment=factor(Dose,levels=unique(Dose)))
dat_medium$Response[dat_medium$Response < 0] <- 0
mod <- drm(Response~Dose,data=dat_medium,fct=LL.3())
fctList <- list(LN.4(),LL.4(),W1.3(),LL2.2())
res <- mselect.plus(mod,fctList = fctList )
modList <- res$modList
p <- plot.modList(modList[1:3])
p
}
}
