% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS_AO.R
\name{plot.StepDownRSCABS}
\alias{plot.StepDownRSCABS}
\title{Plot method for StepDownRSCABS objects}
\usage{
\method{plot}{StepDownRSCABS}(x, ...)
}
\arguments{
\item{x}{A StepDownRSCABS object}

\item{...}{Additional arguments (not used)}
}
\value{
A ggplot object
}
\description{
Plot method for StepDownRSCABS objects
}
