% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS.R
\name{plotRSCABS}
\alias{plotRSCABS}
\title{plotRSCABS}
\usage{
plotRSCABS(
  Data,
  Effect,
  Treatment,
  Metric = "Percent",
  Lowest = "Remove",
  PlotParms = NULL,
  Format = NULL,
  File = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Should be deprecated since the plotting function is very awkward to use.
}
