% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminary.R, R/zzz_aliases.R
\name{prelimSummary}
\alias{prelimSummary}
\alias{prelim_summary}
\title{Preliminary Summary of Dose Response Data}
\usage{
prelimSummary(testdata, dose_col = "Dose", response_col = "Response")

prelim_summary(testdata, dose_col = "Dose", response_col = "Response")
}
\arguments{
\item{testdata}{A data frame containing the dose and response data.}

\item{dose_col}{name of the dose column, default being "Dose".}

\item{response_col}{name of the response column.}
}
\value{
A data frame summarizing the mean, standard deviation, percent inhibition, and coefficient of variation for each dose.
}
\description{
This function calculates the mean response, standard deviation, percent inhibition, and coefficient of variation for each dose level.
}
\examples{
# Sample data frame
testdata <- data.frame(Dose = c(0, 1, 2, 3, 0, 1, 2, 3),
                      Response = c(10, 20, 30, 40, 12, 22, 32, 42))

# Create the summary
summary_result <- prelimSummary(testdata, dose_col = "Dose", response_col = "Response")
print(summary_result)
}
\keyword{PrelimnaryAssessments}
