% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{print.completeTrendAnalysis}
\alias{print.completeTrendAnalysis}
\title{Print method for complete trend analysis (UPDATED)}
\usage{
\method{print}{completeTrendAnalysis}(x, ...)
}
\description{
Print method for complete trend analysis (UPDATED)
}
