% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{print.drcComp}
\alias{print.drcComp}
\title{Print Method for drcComp Objects}
\usage{
\method{print}{drcComp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{drcComp}.}

\item{...}{Additional arguments (currently not used).}
}
\value{
The function prints the \code{Comparison} and \code{EFSA} components
of the object by default.
}
\description{
This function provides a custom print method for objects of class \code{drcComp}.
It prints the \code{Comparison} and \code{EFSA} components of the object.
}
