% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_test.R
\name{print.dunn_test_result}
\alias{print.dunn_test_result}
\title{Print method for dunn_test_result}
\usage{
\method{print}{dunn_test_result}(x, ...)
}
\arguments{
\item{x}{A dunn_test_result object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for dunn_test_result
}
