% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{print.dunnett_test_result}
\alias{print.dunnett_test_result}
\title{Print method for dunnett_test_result objects}
\usage{
\method{print}{dunnett_test_result}(x, ...)
}
\arguments{
\item{x}{A dunnett_test_result object}

\item{...}{Additional arguments passed to print methods}
}
\description{
Print method for dunnett_test_result objects
}
