% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepDownTrendTest_wrapper.R
\name{print.noecTrendTest}
\alias{print.noecTrendTest}
\title{Print method for noecTrendTest objects}
\usage{
\method{print}{noecTrendTest}(x, ...)
}
\arguments{
\item{x}{An object of class \code{noecTrendTest}.}

\item{...}{Other arguments (not used).}
}
\description{
Print method for noecTrendTest objects
}
