% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS_AO.R
\name{print.StepDownRSCABS}
\alias{print.StepDownRSCABS}
\title{Print method for StepDownRSCABS objects}
\usage{
\method{print}{StepDownRSCABS}(x, printLowest = FALSE, ...)
}
\arguments{
\item{x}{A StepDownRSCABS object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the object
}
\description{
Print method for StepDownRSCABS objects
}
