% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{print.stepDownTrendBinom}
\alias{print.stepDownTrendBinom}
\title{Print Method for stepDownTrendBinom Objects}
\usage{
\method{print}{stepDownTrendBinom}(x, ...)
}
\arguments{
\item{x}{An object of class "stepDownTrendBinom"}

\item{...}{Additional arguments passed to print}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for stepDownTrendBinom Objects
}
