% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_binom.R
\name{print.TaroneTrendTest}
\alias{print.TaroneTrendTest}
\title{Print method for TaroneTrendTest}
\usage{
\method{print}{TaroneTrendTest}(x, ...)
}
\arguments{
\item{x}{TaroneTrendTest object}

\item{...}{Additional arguments}
}
\description{
Print method for TaroneTrendTest
}
