% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{pvi_example}
\alias{pvi_example}
\title{Fake data as an example of PVI data}
\format{
collembola_juveniles in wide format
}
\usage{
pvi_example
}
\description{
Fake data as an example of PVI data
}
\keyword{datasets}
