% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{quantal_dat_nested}
\alias{quantal_dat_nested}
\title{Data from Acute Studies}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
quantal_dat_nested
}
\description{
Data are from a study of the response of the daphnids. There are 6 treatment
groups and 6 replicates in each treatment group, 5 individuals in each replicates.
}
\keyword{datasets}
