% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendTest_JG.R
\name{rankTransform}
\alias{rankTransform}
\title{Rank Transform Data Using Blom's Method}
\usage{
rankTransform(Data, VecName)
}
\arguments{
\item{Data}{A data frame containing the variable to be transformed}

\item{VecName}{The name of the variable in the data frame to be transformed}
}
\value{
A data frame with an additional column 'TransformedResponse' containing the rank-transformed values
}
\description{
This function performs a rank transformation on the data using Blom's method,
which is equivalent to the SAS PROC RANK with NORMAL=BLOM and TIES=MEAN options.
The transformation applies the formula (r-3/8)/(n+1/4)
where r is the rank and n is the sample size.
}
\examples{
test_data <- data.frame(Value = c(5, 2, 7, 2, 9, 3))
transformed_data <- rankTransform(test_data, "Value")
}
