% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDD.R
\name{report_dunnett_summary}
\alias{report_dunnett_summary}
\title{Generate a Comprehensive Dunnett Test Summary Report (Robust Version)}
\usage{
report_dunnett_summary(
  formula,
  data,
  dunnett_test_func,
  alternative = "less",
  include_random_effect = FALSE,
  ...
)
}
\arguments{
\item{formula}{The model formula, e.g., Response ~ Dose.}

\item{data}{The dataframe containing the data.}

\item{dunnett_test_func}{The dunnett_test function you use.}

\item{alternative}{The alternative hypothesis to be used for the test.}

\item{...}{Additional arguments passed to your dunnett_test_func.}
}
\value{
A single tibble combining descriptive stats, test results, and MDD\%.
}
\description{
Generate a Comprehensive Dunnett Test Summary Report (Robust Version)
}
