% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_drc_data.R, R/zzz_aliases.R
\name{reshape_drcData}
\alias{reshape_drcData}
\alias{reshape_drc_data}
\title{Reshape the wide data to long data}
\usage{
reshape_drcData(dat, replicate_col = "Replicates")

reshape_drc_data(dat, replicate_col = "Replicates")
}
\arguments{
\item{dat}{data table with Replicate columns and dose groups}

\item{replicate_col}{name of the replicate column or columns need to be retained as identifier columns.}
}
\value{
long format of the dat
}
\description{
Reshape the wide data to long data
}
\examples{
 reshape_drcData(collembola_juveniles) ## note collembola_juveniles is fake data.
}
