% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS.R
\name{runRSCABS}
\alias{runRSCABS}
\title{Run RSCABS test}
\usage{
runRSCABS(Data, Treatment, Replicate = "", Effects = "", test.type = "RS")
}
\arguments{
\item{Data}{A standard data set in the tall format.  Every row indicates an organism.
The data set must contain columns for the treatment level and every tested histological endpoint.}

\item{Treatment}{The name of the column that contains the information about the treatment level.
Increasing values indicate higher treatments.}

\item{Replicate}{The name of the column that contains the information about the replicate structure.
If the  replicate is not specified this will default to running "CA" as the test type.}

\item{Effects}{The endpoint to be tested.  Defaults to all columns that have integers less then 20.
The analysis assumes that higher scores indicate a worse outcome.}

\item{test.type}{Indicate the type of analysis to be performed.  Use "RS" to select the Rao-Scott
adjustment to the Cochran-Armitage test and "CA" to ignore the adjustment.}
}
\value{
a table of the test results for each treatment and injury score.
}
\description{
Runs the Rao-Scott adjusted Cochran-Armitage trend test by slices (RSCABS) analysis.
}
\examples{
\dontrun{
## Not run:
#Take the subset corresponding to F0-females of 16 weeks of age
data(exampleHistData)
exampleHistData.sub<-exampleHistData[which(exampleHistData$Generation=='F2' &
             exampleHistData$Genotypic_Sex=='Female' & exampleHistData$Age=='16_wk' ),  ]
 #Run RSCABS
 eampleResults<-runRSCABS(exampleHistData.sub,'Treatment',
                          'Replicate',test.type='RS')
}
}
\references{
Green, John W. and Springer, Timothy A. and Saulnier, Amy N. and Swintek, Joe,
(2014) Statistical analysis of histopathological endpoints. Environmental Toxicology and Chemistry, 33(5), 1108-1116
}
\author{
Joe Swintek
}
