% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS.R
\name{stepKRSCABS}
\alias{stepKRSCABS}
\title{Steps through the severity score for a given effect}
\usage{
stepKRSCABS(Effect, Data.Prep, Treatment, Replicate, test.type)
}
\arguments{
\item{Effect}{Endpoint being tested.}

\item{Data.Prep}{Data prepared by prepDataRSCABS.}

\item{Treatment}{Name of the treatment variable.}

\item{Replicate}{Name of the replicate variable.}

\item{test.type}{Indicate the type of analysis to be performed.  Use "RS" to select
the Rao-Scott adjustment to the Cochran-Armitage test and "CA" to ignore the adjustment.}
}
\value{
Results.Effect
\itemize{
\item{Results.Effect}{ An intermediary step for results. }
}
}
\description{
An internal function for stepping through each severity score of an endpoint.
}
\author{
Joe Swintek
}
