% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQJT.R
\name{summary.mqjtTest}
\alias{summary.mqjtTest}
\title{Summarise MQJT Results as a Data Frame}
\usage{
\method{summary}{mqjtTest}(object, ...)
}
\arguments{
\item{...}{Additional arguments (ignored).}

\item{x}{An object of class \code{"mqjtTest"}.}
}
\value{
A data frame with one row per cutpoint showing the NOEC and LOEC.
}
\description{
Summarise MQJT Results as a Data Frame
}
