% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSCABS_AO.R
\name{summary.StepDownRSCABS}
\alias{summary.StepDownRSCABS}
\title{Summary method for StepDownRSCABS objects}
\usage{
\method{summary}{StepDownRSCABS}(object, ...)
}
\arguments{
\item{object}{A StepDownRSCABS object}

\item{...}{Additional arguments (not used)}
}
\value{
A list with summary information
}
\description{
Summary method for StepDownRSCABS objects
}
