% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdispersion_binom.R
\name{tarone_with_trend_removal}
\alias{tarone_with_trend_removal}
\title{Internal function to perform Tarone test with trend removal and calculate phi}
\usage{
tarone_with_trend_removal(successes, totals, scores)
}
\arguments{
\item{successes}{Vector of success counts (summarized by dose)}

\item{totals}{Vector of total trials (summarized by dose)}

\item{scores}{Vector of scores defining the trend}
}
\description{
Internal function to perform Tarone test with trend removal and calculate phi
}
