% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{test_cases_data}
\alias{test_cases_data}
\title{test_cases_data for validation purpose}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 768 rows and 17 columns.
}
\usage{
test_cases_data
}
\description{
test_cases_data for validation purpose
}
\keyword{datasets}
