% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepdown_binom.R
\name{test_overdispersion}
\alias{test_overdispersion}
\title{Function to test for overdispersion in binomial data}
\usage{
test_overdispersion(successes, totals, model = NULL)
}
\arguments{
\item{successes}{Numeric vector of successes (e.g., number of survivors).}

\item{totals}{Numeric vector of total observations (e.g., total organisms).}

\item{model}{A fitted binomial GLM with predictors accounted for. (default is NULL)}
}
\value{
a list of dispersion test results
}
\description{
Function to test for overdispersion in binomial data
}
