% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsr_tsk.R
\name{tsk.data.frame}
\alias{tsk.data.frame}
\title{TSK Analysis for Data Frame Input}
\usage{
\method{tsk}{data.frame}(input, control = 0, trim = 0, conf.level = 0.95, use.log.doses = TRUE, ...)
}
\arguments{
\item{input}{A data frame containing the columns \code{x} (doses),
\code{n} (total counts), and \code{r} (response counts).}

\item{control}{A numeric value indicating the control dose (default is 0).}

\item{trim}{A numeric value indicating the trim level (default is 0).}

\item{conf.level}{A numeric value indicating the confidence level
(default is 0.95).}

\item{use.log.doses}{A logical value indicating whether to use
log-transformed doses (default is TRUE).}

\item{...}{Additional arguments passed to the function.}
}
\value{
The result of the TSK analysis.
}
\description{
This function performs TSK analysis for data frame input.
}
