% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsr_tsk.R
\name{tsk}
\alias{tsk}
\title{Trimmed Spearman-Karber Method by brsr}
\usage{
tsk(...)
}
\arguments{
\item{...}{inputs}
}
\value{
tsk estimations
}
\description{
The function is adapted from the single function package tsk developed with
GPL3 license. It is included for validation purpose.
}
\references{
\url{https://github.com/brsr/tsk}
}
