% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R
\name{contEndpoint}
\alias{contEndpoint}
\title{get Endpoint for continuous data}
\usage{
contEndpoint(
  paov,
  pks,
  pnormal,
  phomogeneity,
  monotonicity,
  william,
  dunnett,
  dunn,
  jonckheere,
  procedure = "stepDown",
  doses = c("A", "B", "C", "D")
)
}
\arguments{
\item{doses}{}
}
\description{
get Endpoint for continuous data
}
