% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{drcCompare}
\alias{drcCompare}
\title{Compare models with multiple criteria}
\usage{
drcCompare(
  modRes = NULL,
  modList = NULL,
  trend = "Decrease",
  CI = c("delta", "inv", "bmd-inv"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Compare models with multiple criteria
}
