% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcHelper-package.R
\docType{package}
\name{drcHelper-package}
\alias{drcHelper}
\alias{drcHelper-package}
\title{drcHelper: Collection and verification of helper functions for dose-response analysis}
\description{
Helper functions for drc package and other dose-response analysis functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bayer-group.github.io/drcHelper/index.html}
}

}
\author{
\strong{Maintainer}: Zhenglei Gao \email{zhenglei.gao@bayer.com} (\href{https://orcid.org/0000-0002-4042-310X}{ORCID})

Authors:
\itemize{
  \item Sean Ryan \email{sryan@exponent.com}
}

}
\keyword{internal}
