% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{ED.plus}
\alias{ED.plus}
\title{Calculating ED following the regulatory ED definition.}
\usage{
\method{ED}{plus}(
  object,
  respLev,
  maxEff = TRUE,
  trend = "Increase",
  range = "Percentage",
  CI = c("delta", "inv", "bmd-inv"),
  ...
)
}
\arguments{
\item{object}{fitted model using drc::drm}

\item{respLev}{response level, the x of ECx.}

\item{maxEff}{Not used for now. maximum effect.}

\item{trend}{Increase of Decrease, whether the dose response is decreasing or increasing compared control.}

\item{range}{using Percentage, not used for now.}

\item{...}{other parameters for ED function}

\item{interval}{methods to calculate the confidence intervals!}
}
\value{
Back calculated regulatory ECx
}
\description{
Calculating ED following the regulatory ED definition.
}
\examples{
\dontrun{
}
}
