% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{getEC50}
\alias{getEC50}
\title{Get the EC50 Estimate from a Model}
\usage{
getEC50(mod, approximate = FALSE)
}
\arguments{
\item{mod}{A fitted model object. This can be of class "glm" or "glmmPQL",
or other models compatible with the \code{ED} function.}

\item{approximate}{logical, passing into backCalcSE}
}
\value{
A data frame containing the following columns:
\item{EC50}{The estimated EC50 value.}
\item{lower}{The lower bound of the 95\% confidence interval for the EC50.}
\item{upper}{The upper bound of the 95\% confidence interval for the EC50.}
\item{se}{The standard error of the EC50 estimate.}
}
\description{
This function calculates the EC50 (the dose at which 50\% of the maximum effect is observed)
from a fitted model. It supports generalized linear models (GLMs) and generalized linear mixed models (GLMMs).
}
\examples{
# Assuming `fit` is a fitted glm model
fit <- glm(yt ~ log(dose),family= quasibinomial(link="logit"),data=pvi_example)
ec50_result <- getEC50(fit)
print(ec50_result)

}
