% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{getModelName}
\alias{getModelName}
\title{Get Model Name}
\usage{
getModelName(fname = NULL)
}
\arguments{
\item{fname}{a character string contains the name of the fitted model.}
}
\value{
the model name
}
\description{
Get Model Name
}
\examples{
getModelName("LL.2")

}
