% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R
\name{getwilliamRes}
\alias{getwilliamRes}
\title{get from william res accept/reject}
\usage{
getwilliamRes(william)
}
\arguments{
\item{william}{william test results}
}
\description{
get from william res accept/reject
}
