% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brsr_tsk.R, R/data_description.R
\docType{data}
\name{hamilton}
\alias{hamilton}
\title{This is hamilton data}
\format{
A list containing ten data frames: dr1a, dr1b, dr1c,
dr1d, dr1e, dr4a, dr4b, dr4c, dr4d, dr4e,

A list containing ten data frames: dr1a, dr1b, dr1c,
dr1d, dr1e, dr4a, dr4b, dr4c, dr4d, dr4e
}
\source{
Hamilton, 1977.

Hamilton, 1977.
}
\usage{
hamilton

hamilton
}
\description{
Example dose-response data given in Hamilton (1977).
Note that, as per Hamilton (1978), the confidence intervals
given in Hamilton (1977) for these data sets are incorrect.

Example dose-response data given in Hamilton (1977).
Note that, as per Hamilton (1978), the confidence intervals
given in Hamilton (1977) for these data sets are incorrect.
}
\references{
\url{https://github.com/brsr/tsk}

\url{https://github.com/brsr/tsk}
}
\author{
B R S Recht
}
\keyword{data}
\keyword{datasets}
