% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendTest_JG.R
\name{monotonicityTest}
\alias{monotonicityTest}
\title{Testing Monotonicity}
\usage{
monotonicityTest(Data, Treatment, Response)
}
\arguments{
\item{Data}{Data frame}

\item{Treatment}{name of the treatment variable}

\item{Response}{name of the response variable}
}
\value{
monotonicity table
}
\description{
Testing Monotonicity
}
\examples{
\dontrun{
x <- c(106, 114, 116, 127, 145,110, 125, 143, 148, 151,
136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("0", "I", "II")
monotonicityTest(data.frame(treatment_var = g,response_var=x), "treatment_var", "response_var")
mock_data <- data.frame(
treatment_var = factor(rep(c("Control", "Dose1", "Dose2", "Dose3"), each = 10)),
response_var = c(rnorm(10, mean = 5), rnorm(10, mean = 7), rnorm(10, mean = 8), rnorm(10, mean = 10))
)
monotonicityTest(mock_data, "treatment_var", "response_var")
}
}
