% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrendTest_JG.R
\name{monotonicityTest}
\alias{monotonicityTest}
\title{Testing Monotonicity}
\usage{
monotonicityTest(Data, Treatment, Response)
}
\arguments{
\item{Data}{Data frame}

\item{Treatment}{name of the treatment variable}

\item{Response}{name of the response variable}
}
\value{
monotonicity table
}
\description{
Testing Monotonicity
}
