% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{mselect.ED}
\alias{mselect.ED}
\title{Select ECx estimation from Models with mselect.plus and EFSA SO criteria}
\usage{
mselect.ED(modList, respLev = c(10, 20, 50), trend = "Decrease", ...)
}
\arguments{
\item{...}{}
}
\description{
Select ECx estimation from Models with mselect.plus and EFSA SO criteria
}
