% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{mselect.ED}
\alias{mselect.ED}
\title{ECx calculation together with normalized width proposed by EFSA SO.}
\usage{
mselect.ED(modList, respLev = c(10, 20, 50), trend = "Decrease", ...)
}
\arguments{
\item{modList}{list of models}

\item{...}{}
}
\value{
ED result table
}
\description{
ECx calculation together with normalized width proposed by EFSA SO.
}
