% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drc_Helper.R
\name{mselect.plus}
\alias{mselect.plus}
\title{added functionality for mselect}
\usage{
mselect.plus(
  object = NULL,
  fctList = NULL,
  nested = FALSE,
  sorted = c("IC", "Res var", "Lack of fit", "no"),
  linreg = FALSE,
  icfct = AIC,
  respCol = "effect",
  doseCol = "dose",
  data = NULL,
  type = "continuous",
  additionalReliability = c("EFSA")
)
}
\arguments{
\item{type}{}

\item{additionalReliability}{additional reliability need to be calculated}
}
\value{
mselect.plus
}
\description{
added functionality for mselect
}
