% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R
\name{pavaMean}
\alias{pavaMean}
\title{Calculate Means Using the Pool Adjacent Violators Algorithm (PAVA)}
\usage{
pavaMean(x, g, alternative = "greater")
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{g}{A factor variable that defines the groups for which to calculate means.}

\item{alternative}{A character string specifying the alternative hypothesis.
Options are "greater" (default) or "less".}
}
\value{
A data frame containing the following columns:
\item{pavaMean}{The adjusted means for each group.}
\item{SE.diff}{The standard errors of the differences.}
}
\description{
This function computes adjusted means and standard errors for groups defined by a factor variable
using the Pool Adjacent Violators Algorithm (PAVA).
}
\examples{
# Example usage:
x <- c(1, 2, 3, 4, 5, 6)
g <- factor(c("A", "A", "B", "B", "C", "C"))
result <- pavaMean(x, g)
print(result)
x <- c(106, 114, 116, 127, 145,
       110, 125, 143, 148, 151,
       136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("0", "I", "II")
pavaMean(x,g)
}
