% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelimnary.R
\name{prelimPlot1}
\alias{prelimPlot1}
\title{Preliminary Plot 1 for Dose Response Data}
\usage{
prelimPlot1(
  testdata,
  ylab = "Response",
  xlab = "Test Concentration [nominal, mg a.s./L]",
  title = "Measured Variable"
)
}
\arguments{
\item{testdata}{A data frame containing the dose and response data.}

\item{ylab}{A string for the y-axis label. Default is "Response".}

\item{xlab}{A string for the x-axis label. Default is "Test Concentration \link{nominal, mg a.s./L}".}

\item{title}{A string for the plot title. Default is "Measured Variable".}
}
\value{
A ggplot object.
}
\description{
This function generates a scatter plot of response against nominal dose.
}
