% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelimnary.R
\name{prelimPlot3}
\alias{prelimPlot3}
\title{Preliminary Plot 3 for Dose Response Data}
\usage{
prelimPlot3(
  testdata,
  ylab = "Response",
  xlab = "Test Concentration [nominal, mg a.s./L]",
  title = "Measured Variable",
  a = 1.96,
  dose_col = "Dose",
  response_col = "Response"
)
}
\arguments{
\item{testdata}{A data frame containing the dose and response data.}

\item{ylab}{A string for the y-axis label. Default is "Response".}

\item{xlab}{A string for the x-axis label. Default is \code{"Test Concentration [nominal, mg a.s./L]"}.}

\item{title}{A string for the plot title. Default is "Measured Variable".}

\item{a}{the quantile for corresponding CI for mean. default is qnorm(0.975).}

\item{response_col}{name of the response column. description}

\item{dosecol}{name of the dose column, default being "Dose".  description}
}
\value{
A ggplot object.
}
\description{
This function generates a scatter plot of response against nominal dose.
}
\keyword{PrelimnaryAssessments}
