% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prelimnary.R
\name{prelimSummary}
\alias{prelimSummary}
\title{Preliminary Summary of Dose Response Data}
\usage{
prelimSummary(testdata)
}
\arguments{
\item{testdata}{A data frame containing the dose and response data.}
}
\value{
A data frame summarizing the mean, standard deviation, percent inhibition, and coefficient of variation for each dose.
}
\description{
This function calculates the mean response, standard deviation, percent inhibition, and coefficient of variation for each dose level.
}
