% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_DRC_Data.R
\name{reshape_drcData}
\alias{reshape_drcData}
\title{Reshape the wide data to long data}
\usage{
reshape_drcData(dat)
}
\arguments{
\item{dat}{data table with Replicate columns and dose groups}
}
\value{
long format of the dat
}
\description{
Reshape the wide data to long data
}
\examples{
 reshape_drcData(collembola_juveniles)
}
