% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Helper.R
\name{simplifyTreatment}
\alias{simplifyTreatment}
\title{Resolve the excel datasheet number issue like 12.300000000000001 or 8.3579999999999}
\usage{
simplifyTreatment(trt)
}
\arguments{
\item{trt}{a treatment vector, either factor or character}
}
\value{
a simplified vector
}
\description{
Resolve the excel datasheet number issue like 12.300000000000001 or 8.3579999999999
}
\examples{
x <- structure(c(1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L, 3L, 3L, 3L,
3L, 4L, 4L, 4L, 4L, 5L, 5L, 5L, 5L, 6L, 6L, 6L, 6L, 1L, 1L, 1L,
1L, 1L, 1L, 2L, 2L, 2L, 2L, 3L, 3L, 3L, 3L, 4L, 4L, 4L, 4L, 5L,
5L, 5L, 5L, 6L, 6L, 6L, 6L, 1L, 1L, 1L, 1L, 1L, 1L, 2L, 2L, 2L,
2L, 3L, 3L, 3L, 3L, 4L, 4L, 4L, 4L, 5L, 5L, 5L, 5L, 6L, 6L, 6L,
6L), levels = c("Control", "1", "3.51", "12.300000000000001",
             "42.9", "150"), class = "factor")
simplifyTreatment(x)
}
