% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R
\name{summaryZG}
\alias{summaryZG}
\title{Summary Williams Test results.}
\usage{
summaryZG(object, ...)
}
\arguments{
\item{object}{William test result object}

\item{...}{}
}
\value{
William test results
}
\description{
Summary Williams Test results.
}
