% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Endpoints.R
\name{summaryZG}
\alias{summaryZG}
\title{Summary Williams Test results.}
\usage{
summaryZG(object, verbose = F, ...)
}
\arguments{
\item{object}{William test result object}

\item{...}{additional parameters to be passed into the function, placeholder}
}
\value{
William test results
}
\description{
Summary Williams Test results.
}
\examples{
## Example from Sachs (1997, p. 402)
x <- c(106, 114, 116, 127, 145,
       110, 125, 143, 148, 151,
       136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("0", "I", "II")

## Williams Test
res <- williamsTest(x ~ g)
summaryZG(res) ## return a data frame instead of a list.
}
