% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_DRC_Data.R
\name{treatment2dose}
\alias{treatment2dose}
\title{Change Treatment groups to numerical dose}
\usage{
treatment2dose(x)
}
\arguments{
\item{x}{treatment groups with numbers and "Control"}
}
\value{
numeric vector of dose
}
\description{
Change Treatment groups to numerical dose
}
