% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Helper.R
\name{wide2long}
\alias{wide2long}
\title{change from wide format to long format}
\usage{
wide2long(widedat, cnames = 1, repnames = 1)
}
\arguments{
\item{cnames}{}
}
\description{
change from wide format to long format
}
