% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCharrms.R
\name{williamsTest_JG}
\alias{williamsTest_JG}
\title{Williams Test from the StatCharrms Package}
\usage{
williamsTest_JG(df, resp, trt, direction = "decreasing", SeIn = NULL)
}
\arguments{
\item{df}{The data frame}

\item{resp}{The name of the response as a string}

\item{trt}{The name of the response as a string}

\item{direction}{Is the direction of the test 'decreasing' or 'increasing'}

\item{SeIn}{the standard error, default is program selected. WilliamsTest can take in a different value in the case of repeated measures}
}
\value{
Williams' test result
}
\description{
The function is adapted from the archived version of StatCharrms developed by
Joe Swintek et al with CC0 license. It is not updated anymore and included
for validation purpose. We recommend to use the williamsTest from PMCMRplus
package instead.
}
\examples{
## Williams Test
x <- c(106, 114, 116, 127, 145,110, 125, 143, 148, 151,
136, 139, 149, 160, 174)
g <- gl(3,5)
levels(g) <- c("0", "I", "II")
PMCMRplus::williamsTest(x ~ g)
williamsTest_JG(data.frame(treatment_var = g,response_var=x),"response_var","treatment_var",direction="increasing")
}
