% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCharrms.R
\name{williamsTest_JG}
\alias{williamsTest_JG}
\title{Williams Test from the StatCharrms Package}
\usage{
williamsTest_JG(df, resp, trt, direction = "decreasing", SeIn = NULL)
}
\arguments{
\item{SeIn}{}
}
\description{
Williams Test from the StatCharrms Package
}
